+++ src/main/kotlin/blackjack/domain/card/CardTier.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardTier(
     val values: Int,
