+++ src/main/kotlin/blackjack/domain/deck/Deck.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.deck
 
-import blackjack.domain.TrumpCard
+import blackjack.domain.card.TrumpCard
 
 interface Deck {
     fun pop(): TrumpCard
