+++ src/main/kotlin/blackjack/domain/participant/Participant.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.participant
 
-import blackjack.domain.TrumpCard
+import blackjack.domain.card.TrumpCard
 
 sealed interface Participant {
     val cards: List<TrumpCard>
