+++ src/main/kotlin/blackjack/dsl/skill/SkillBuilder.kt
@@ -4,11 +4,11 @@ class SkillBuilder {
     private lateinit var soft: String
     private lateinit var hard: String
 
-    fun soft(value: String)  {
+    fun soft(value: String) {
         soft = value
     }
 
-    fun hard(value: String)  {
+    fun hard(value: String) {
         hard = value
     }
