+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,9 +1,9 @@
 package blackjack.view
 
-import blackjack.domain.CardTier
 import blackjack.domain.GameResult
-import blackjack.domain.Shape
-import blackjack.domain.TrumpCard
+import blackjack.domain.card.CardTier
+import blackjack.domain.card.Shape
+import blackjack.domain.card.TrumpCard
 import blackjack.domain.participant.Dealer
 import blackjack.domain.participant.Player
 import java.lang.String.format
