+++ src/test/kotlin/blackjack/BlackJackGameTest.kt
@@ -1,10 +1,10 @@
 package blackjack
 
 import blackjack.domain.BlackJackGame
-import blackjack.domain.CardTier
-import blackjack.domain.Shape
-import blackjack.domain.TrumpCard
 import blackjack.domain.UserChoice
+import blackjack.domain.card.CardTier
+import blackjack.domain.card.Shape
+import blackjack.domain.card.TrumpCard
 import blackjack.domain.deck.ShuffledDeck
 import blackjack.fixture.playersFixture
 import org.assertj.core.api.Assertions.assertThat
