+++ src/test/kotlin/blackjack/OrderedDeck.kt
@@ -1,12 +1,13 @@
-package blackjack.domain.deck
+package blackjack
 
 import blackjack.domain.card.CardTier
 import blackjack.domain.card.Shape
 import blackjack.domain.card.TrumpCard
+import blackjack.domain.deck.Deck
 import java.util.ArrayDeque
 import java.util.Deque
 
-class ShuffledDeck : Deck {
+class OrderedDeck : Deck {
     private val cards: Deque<TrumpCard> = ArrayDeque()
 
     init {
@@ -21,5 +22,5 @@ class ShuffledDeck : Deck {
                 CardTier.entries.map { tier ->
                     TrumpCard(tier, shape)
                 }
-            }.shuffled()
+            }
 }
